Refactor old ReportGenerator
