+++ subprojects/internal-performance-testing/src/main/groovy/org/gradle/performance/results/CrossVersionResultsStore.java
@@ -51,6 +51,16 @@ import static org.gradle.performance.results.ResultsStoreHelper.toArray;
  * A {@link DataReporter} implementation that stores results in an H2 relational database.
  */
 public class CrossVersionResultsStore implements DataReporter<CrossVersionPerformanceResults>, ResultsStore {
+    private static final String FLAKINESS_RATE_SQL =
+        "SELECT TESTID, AVG(CONVERT(CASEWHEN(DIFFCONFIDENCE > 0.99, 1, 0), DECIMAL)) AS FAILURE_RATE\n" +
+            "FROM TESTEXECUTION\n" +
+            "WHERE (CHANNEL = 'flakiness-detection-master' OR CHANNEL = 'flakiness-detection-release')\n" +
+            "GROUP BY TESTID";
+    private static final String FAILURE_THRESOLD_SQL =
+        "SELECT TESTID, MAX(ABS((BASELINEMEDIAN-CURRENTMEDIAN)/BASELINEMEDIAN)) as THRESHOLD\n" +
+            "FROM TESTEXECUTION\n" +
+            "WHERE (CHANNEL = 'flakiness-detection-master' or CHANNEL= 'flakiness-detection-release') AND DIFFCONFIDENCE > 0.99\n" +
+            "GROUP BY TESTID";
     private final long ignoreV17Before;
     private final PerformanceDatabase db;
     private final Map<String, GradleVersion> gradleVersionCache = new HashMap<>();
@@ -380,11 +390,11 @@ public class CrossVersionResultsStore implements DataReporter<CrossVersionPerfor
     }
 
     public Map<String, BigDecimal> getFlakinessRates() {
-        return queryFlakinessData(PerformanceFlakinessDataProvider.FLAKINESS_RATE_SQL);
+        return queryFlakinessData(FLAKINESS_RATE_SQL);
     }
 
     public Map<String, BigDecimal> getFailureThresholds() {
-        return queryFlakinessData(PerformanceFlakinessDataProvider.FAILURE_THRESOLD_SQL);
+        return queryFlakinessData(FAILURE_THRESOLD_SQL);
     }
 
     private Map<String, BigDecimal> queryFlakinessData(String sql) {
