+++ subprojects/internal-performance-testing/src/main/groovy/org/gradle/performance/results/report/HtmlPageGenerator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.performance.results;
+package org.gradle.performance.results.report;
 
 import com.google.common.base.Charsets;
 import com.google.common.base.Joiner;
@@ -23,6 +23,7 @@ import com.googlecode.jatl.Html;
 import org.gradle.performance.measure.Amount;
 import org.gradle.performance.measure.DataSeries;
 import org.gradle.performance.measure.Duration;
+import org.gradle.performance.results.MeasuredOperationList;
 import org.gradle.reporting.ReportRenderer;
 import org.gradle.util.GradleVersion;
