+++ subprojects/internal-performance-testing/src/main/groovy/org/gradle/performance/results/report/TestDataGenerator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,16 @@
  * limitations under the License.
  */
 
-package org.gradle.performance.results;
+package org.gradle.performance.results.report;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import groovy.json.JsonGenerator;
+import org.gradle.performance.results.CrossVersionPerformanceTestHistory;
+import org.gradle.performance.results.FormatSupport;
+import org.gradle.performance.results.MeasuredOperationList;
+import org.gradle.performance.results.PerformanceTestExecution;
+import org.gradle.performance.results.PerformanceTestHistory;
 import org.gradle.reporting.ReportRenderer;
 
 import java.io.IOException;
