+++ subprojects/internal-performance-testing/src/test/groovy/org/gradle/performance/results/report/ReportGeneratorTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package org.gradle.performance.results
+package org.gradle.performance.results.report
 
 import org.gradle.performance.ResultSpecification
+import org.gradle.performance.results.CrossVersionResultsStore
+import org.gradle.performance.results.ResultsStore
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
 import org.gradle.util.SetSystemProperties
 import org.junit.Rule
-
-import static org.gradle.performance.results.PerformanceFlakinessDataProvider.EmptyPerformanceFlakinessDataProvider.*
 
 class ReportGeneratorTest extends ResultSpecification {
     @Rule
@@ -51,7 +51,7 @@ class ReportGeneratorTest extends ResultSpecification {
         store.report(result2)
 
         when:
-        generator.generate(reportDir, resultsJson, 'testProject')
+        generator.generateReport(reportDir.absolutePath, resultsJson.absolutePath, 'testProject')
 
         then:
         reportDir.file("index.html").isFile()
